requirements.txt
python-dotenv==1.0.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-openai>=0.1.0
google-search-results==2.4.2
openai>=1.10.0